CIF-2144 - Change price range filters to Double

+++ extensions/product-recs/bundle/src/test/java/com/adobe/cq/commerce/extensions/recommendations/internal/models/v1/common/PriceRangeImplTest.java
@@ -23,9 +23,9 @@ public class PriceRangeImplTest {
 
     @Test
     public void testGetters() {
-        PriceRange priceRange = new PriceRangeImpl(10L, 12L);
-        assertEquals(Long.valueOf(10), priceRange.getMinPrice());
-        assertEquals(Long.valueOf(12), priceRange.getMaxPrice());
+        PriceRange priceRange = new PriceRangeImpl(10.0, 12.0);
+        assertEquals(Double.valueOf(10), priceRange.getMinPrice());
+        assertEquals(Double.valueOf(12), priceRange.getMaxPrice());
     }
 
 }
